refactor: separate query parameter schemas into module-specific files
