+++ typescript/src/plugins/core-account-query-plugin/tools/queries/get-account-query.ts
@@ -2,10 +2,10 @@ import { z } from 'zod';
 import { Client } from '@hashgraph/sdk';
 import { Context } from '@/shared/configuration';
 import { getMirrornodeService } from '@/shared/hedera-utils/mirrornode/hedera-mirrornode-utils';
-import { accountQueryParameters } from '@/shared/parameter-schemas/query.zod';
 import { Tool } from '@/shared/tools';
 import { PromptGenerator } from '@/shared/utils/prompt-generator';
 import { AccountResponse } from '@/shared/hedera-utils/mirrornode/types';
+import { accountQueryParameters } from '@/shared/parameter-schemas/account.zod';
 
 export const getAccountQueryPrompt = (context: Context = {}) => {
   const contextSnippet = PromptGenerator.getContextSnippet(context);
@@ -43,11 +43,12 @@ export const getAccountQuery = async (
       humanMessage: postProcess(account),
     };
   } catch (error) {
-    console.error('Error getting account query', error);
-    if (error instanceof Error) {
-      return error.message;
-    }
-    return 'Failed to get account query';
+    const message = error instanceof Error ? error.message : 'Error getting account query';
+
+    return {
+      raw: { accountId: params.accountId, error: message },
+      humanMessage: message,
+    };
   }
 };
