+++ typescript/src/plugins/core-queries-plugin/tools/queries/get-account-query.ts
@@ -2,10 +2,10 @@ import { z } from 'zod';
 import { Client } from '@hashgraph/sdk';
 import { Context } from '@/shared/configuration';
 import { getMirrornodeService } from '@/shared/hedera-utils/mirrornode/hedera-mirrornode-utils';
-import { accountQueryParameters } from '@/shared/parameter-schemas/query.zod';
 import { Tool } from '@/shared/tools';
 import { PromptGenerator } from '@/shared/utils/prompt-generator';
 import { AccountResponse } from '@/shared/hedera-utils/mirrornode/types';
+import { accountQueryParameters } from '@/shared/parameter-schemas/account.zod';
 
 export const getAccountQueryPrompt = (context: Context = {}) => {
   const contextSnippet = PromptGenerator.getContextSnippet(context);
