+++ typescript/src/plugins/core-token-query-plugin/tools/queries/get-token-info-query.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 import { Context } from '@/shared/configuration';
 import { getMirrornodeService } from '@/shared/hedera-utils/mirrornode/hedera-mirrornode-utils';
-import { tokenInfoQueryParameters } from '@/shared/parameter-schemas/query.zod';
+import { tokenInfoQueryParameters } from '@/shared/parameter-schemas/token.zod';
 import { Client } from '@hashgraph/sdk';
 import { Tool } from '@/shared/tools';
 import { PromptGenerator } from '@/shared/utils/prompt-generator';
