+++ typescript/src/shared/parameter-schemas/token.zod.ts
@@ -126,3 +126,8 @@ export const deleteTokenParameters = (_context: Context = {}) =>
 
 export const deleteTokenParametersNormalised = (_context: Context = {}) =>
   deleteTokenParameters(_context).extend({});
+
+export const tokenInfoQueryParameters = (_context: Context = {}) =>
+  z.object({
+    tokenId: z.string().optional().describe('The token ID to query.'),
+  });
